marketing-assistant-v2/services/customer-analyst/requirements.txt
a2a-sdk[http-server]>=0.3.25
httpx>=0.27.0
pydantic>=2.0.0
uvicorn>=0.27.0
